moved maps folder to  folder
