+++ sites/geohub/src/components/maps/MapStyleCardList.svelte
